Finally Stable!

+++ domain/src/main/java/com/example/domain/models/requests/HabitDoneBody.kt
@@ -4,5 +4,5 @@ import com.google.gson.annotations.SerializedName
 
 data class HabitDoneBody(
     @SerializedName("habit_uid") val uid: String,
-    val date: Long,
+    val date: Int,
 )
